+++ src/logic/Errors/friendly.ts
@@ -105,7 +105,7 @@ export function convertHelpfullError( inputs: IHelpfullInput ) : IHelpfullOutput
   if (result.indexOf('Failed to fetch') > -1 ) { friendlyMessage = 'This can happen if the web url is not valid.'; }
   if (result.indexOf('A null value was detected in the items of a collection property value') > -1 ) { friendlyMessage= 'This can happen if you are saving a null value where an array is expected... Maybe try saving an empty array instead :).'; }
 
-  if (result.indexOf('An unexpected \'StartObject\' node was found when reading from the JSON reader. A \'PrimitiveValue\' node was expected.') > -1 ) { 
+  if (result.indexOf(`An unexpected 'StartObject' node was found when reading from the JSON reader. A 'PrimitiveValue' node was expected.`) > -1 ) { 
     friendlyMessage = 'Common causes:  Saving a Url Object or JSON Object to string column.';
   }
 
@@ -116,7 +116,7 @@ export function convertHelpfullError( inputs: IHelpfullInput ) : IHelpfullOutput
     friendlyMessage = 'Common causes:  The formula refers to the current column which creates circular reference.';
   }
 
-  if (result.indexOf('An unexpected \'PrimitiveValue\' node was found when reading from the JSON reader. A \'StartObject\' node was expected') > -1 ) { 
+  if (result.indexOf(`An unexpected 'PrimitiveValue' node was found when reading from the JSON reader. A 'StartObject' node was expected`) > -1 ) { 
     friendlyMessage = 'Common causes:  Saving a string to a URL column, saving text to multi-select choice column.';
   }
   if (result.indexOf('does not exist') > -1 && result.indexOf('Column') > -1) {